refactor: move render process files to "render" folder
