+++ src/render/Telas/Config/Velocity.js
